requirements.txt
asgiref==3.5.2
Django~=4.1.3
djangorestframework~=3.14.0
djangorestframework-simplejwt==5.2.2
psycopg2==2.9.5
PyJWT==1.7.1
python-dotenv~=0.21.0
pytz==2022.6
sqlparse==0.4.3
tzdata==2022.6

requests~=2.28.1
urllib3~=1.26.13
djangorestframework-jwt~=1.11.0
